fix detekt
